+++ src/Controller/ShopController.php
@@ -10,16 +10,17 @@
 
 namespace c975L\ShopBundle\Controller;
 
+use c975L\ShopBundle\Service\ShopServiceInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use c975L\ShopBundle\Service\ShopServiceInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class ShopController extends AbstractController
 {
-    public function __construct(private readonly ShopServiceInterface $shopService)
-    {
+    public function __construct(
+        private readonly ShopServiceInterface $shopService,
+    ) {
     }
 
     // INDEX